Installed: mock-knex; Added: using mock knex to run unit test;
