+++ server/models/TrustModel.spec.js
@@ -1,9 +1,26 @@
 const TrustModel = require("./TrustModel");
 const {expect} = require("chai");
+const mockKnex = require("mock-knex");
+const tracker = mockKnex.getTracker();
+const knex = require("../database/knex");
 
 describe("TrustModel", () => {
 
+  before(() => {
+    mockKnex.mock(knex);
+    tracker.install();
+  });
+
+  after(() => {
+    mockKnex.unmock(knex);
+  });
+
   it("get trust_relationships", async () => {
+    tracker.on("query", (query) => {
+      query.response([{
+        a:1,
+      }]);
+    });
     const trustModel = new TrustModel();
     const trust_relationships = await trustModel.get();
     expect(trust_relationships).lengthOf(1);
